refactor: use @scalar/galaxy inside @scalar/api-reference
