+++ examples/cdn-api-reference/vite.config.ts
@@ -9,7 +9,7 @@ export default defineConfig({
         // Resolve the uncompiled source code for all @scalar packages
         // @scalar/* -> packages/*/
         // (not @scalar/*/style.css)
-        find: /^@scalar\/(?!(openapi-parser|snippetz|components\/style\.css|components\b))(.+)/,
+        find: /^@scalar\/(?!(openapi-parser|snippetz|galaxy|components\/style\.css|components\b))(.+)/,
         replacement: path.resolve(__dirname, '../../packages/$2/src/index.ts'),
       },
     ],
