+++ packages/api-client-proxy/vite.config.ts
@@ -29,7 +29,7 @@ export default defineConfig({
         // Resolve the uncompiled source code for all @scalar packages
         // @scalar/* -> packages/*/
         // (not @scalar/*/style.css)
-        find: /^@scalar\/(?!(snippetz|components\/style\.css|components\b))(.+)/,
+        find: /^@scalar\/(?!(snippetz|galaxy|components\/style\.css|components\b))(.+)/,
         replacement: path.resolve(__dirname, '../$2/src/index.ts'),
       },
     ],
