+++ packages/galaxy/src/specifications/3.1.yaml
@@ -60,7 +60,7 @@ security:
 tags:
   - name: Authentication
     description:
-      Some endpoints are public, but some required authentication. We provide
+      Some endpoints are public, but some require authentication. We provide
       all the required endpoints to create an account and authorize yourself.
   - name: Planets
     description: Everything about planets
