--- packages/mock-server/src/lib/index.ts
@@ -1,1 +1,0 @@
-export * from './createMockServer'
