+++ packages/mock-server/tests/galaxy.test.ts
@@ -1,7 +1,7 @@
-import { createMockServer } from '@scalar/mock-server'
+import galaxy from '@scalar/galaxy/3.1.yaml?raw'
 import { describe, expect, it } from 'vitest'
 
-import galaxy from '../src/specifications/3.1.yaml?raw'
+import { createMockServer } from '../src/createMockServer'
 
 describe('createMockServer', () => {
   it('GET /planets -> example JSON', async () => {
